requirements.txt
composio-core>=0.7.0
httpx>=0.27.0
